seperated the real html from the the table Proof of Concept
