+++ src/scripts/main.ts
@@ -1,6 +1,8 @@
 import '../styles/style.css';
 
 import { CalendarElement } from '../components/calendar.element';
+import { createYear } from '../components/calendar.table.elements';
+import { Calendar } from './calendar';
 
 const { BASE_URL } = ( import.meta as any ).env;
 
@@ -10,12 +12,12 @@ document.addEventListener( 'DOMContentLoaded', () =>
 {
   const main = document.querySelector( 'main' )
 
-  //const calendar: Calendar = new Calendar( 2024 )
-  const calendar_element = new CalendarElement( 2024 );
+  const calendar_element = new CalendarElement( 2024 )
+  const calendar_table = new Calendar( 2024 )
 
   if ( main )
   {
-    //main.appendChild( createYear( calendar.days ) )
     main.appendChild( calendar_element );
+    main.appendChild( createYear( calendar_table.days ) );
   }
 } )
